scrapers/requirements.txt
# Core Scraping Framework
scrapy==2.11.0
scrapy-playwright==0.0.34

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Async Support
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
supabase==2.3.0

# Proxy & Anti-Detection
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
fake-useragent==1.4.0

# Workflow Orchestration
prefect==2.14.11

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

